move it to pkg/dblock instead
